pis_ofap/requirements.txt
appdirs==1.4.4
asgiref==3.2.7
astroid==2.4.2
certifi==2019.9.11
cffi==1.13.2
chardet==3.0.4
colorama==0.4.3
coreapi==2.3.3
coreschema==0.0.4
cryptography==2.8
distlib==0.3.0
Django==2.2.9
django-cors-headers==3.3.0
django-oauth-toolkit==1.2.0
djangorestframework==3.10.3
drf-yasg==1.17.0
filelock==3.0.12
idna==2.8
inflection==0.3.1
isort==4.3.21
itypes==1.1.0
Jinja2==2.10.3
lazy-object-proxy==1.4.3
MarkupSafe==1.1.1
mccabe==0.6.1
oauthlib==3.1.0
openapi-codec==1.3.2
packaging==19.2
pipenv==2020.6.2
pycparser==2.19
pylint==2.5.3
pyparsing==2.4.5
pytz==2019.3
requests==2.22.0
ruamel.yaml==0.16.5
ruamel.yaml.clib==0.2.0
simplejson==3.17.0
six==1.13.0
sqlparse==0.3.0
toml==0.10.1
uritemplate==3.0.0
urllib3==1.25.7
virtualenv==20.0.23
virtualenv-clone==0.5.4
wrapt==1.12.1